+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## HEAD
+
+* Add compatibility with calendars other than `Calendar.ISO`.
+
 ## v0.1.0 (2019-01-26)
 
 * First public release